dashboard/requirements.txt
streamlit
mysql-connector-python
requests